step 1 in applying requested changes
